requirements/base.txt
wheel==0.37.0
django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
requests==2.28.0

celery==5.2.1
watchdog==2.1.6
PyYAML==6.0
django-celery-results==2.2.0
django-celery-beat==2.2.1
